refactor pagination component
