+++ design-at-cornell/src/explore-courses/ExploreCourses.tsx
@@ -3,7 +3,7 @@ import axios from 'axios';
 import Title from './title/Title';
 import { VerticalFlex } from '../components/ContainerStyles';
 import Courses from './courses/Courses';
-import Pagination from './Pagination';
+import Pagination from '../pagination/Pagination';
 import { Course } from '../../../server/types';
 import {
   Filters,
@@ -68,8 +68,8 @@ const ExploreCourses = () => {
       <Courses {...displayedCourses} />
       <Pagination
         currentPage={currentPage}
-        coursesPerPage={coursesPerPage}
-        totalCourses={filterResult.length}
+        cardsPerPage={coursesPerPage}
+        totalCards={filterResult.length}
         paginate={setPage}
       />
     </VerticalFlex>
